IGNITE-23554 ScanQuery code cleanup (#11645)
