+++ modules/core/src/main/java/org/apache/ignite/internal/processors/query/GridQueryProcessor.java
@@ -349,8 +349,7 @@ public class GridQueryProcessor extends GridProcessorAdapter {
         initQueryEngines();
 
         idxBuildStatusStorage = new IndexBuildStatusStorage(ctx);
-
-        txAwareQueriesEnabled = ctx.config().getTransactionConfiguration().isTxAwareQueriesEnabled();
+        txAwareQueriesEnabled = U.isTxAwareQueriesEnabled(ctx);
     }
 
     /** {@inheritDoc} */
